+++ tests/unit/test_configuration.py
@@ -1,0 +1,5 @@
+from crawlee.configuration import Configuration
+
+
+def test_global_configuration_works() -> None:
+    assert Configuration.get_global_configuration() is Configuration.get_global_configuration()